Adds signup controller
